completed 1-2 by creating new components and refactoring the app component
